pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "simplefold"
version = "0.1.0"
description = "Folding proteins with SimpleFold."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [
    { name = "Yuyang Wang", email = "yuyangw@apple.com" },
    { name = "Miguel Angel Bautista Martin", email = "mbautistamartin@apple.com" }
]
dependencies = [
    "lightning==2.5.2",
    "biopython==1.85",
    "hydra-colorlog==1.2.0",
    "mediapy==1.0.2",
    "ipykernel==6.22.0",
    "ipyvtklink==0.2.3",
    "gemmi==0.7.3",
    "biotite==1.2.0",
    "rdkit==2025.3.5",
    "gemmi==0.7.3",
    "p-tqdm==1.4.2",
    "einops==0.8.1",
    "mashumaro==3.16",
    "ihm==2.7",
    "modelcif==1.4",
    "dm-tree==0.1.9",
    "click==8.2.1",
    "timm==1.0.19",
    "py3Dmol==2.5.2",
    "scikit-learn==1.7.1",
    "pandas==2.3.1",
    "seaborn==0.13.2",
    "jax[cuda13]",
    "orbax",
    "flax",
    "fair-esm @ git+https://github.com/facebookresearch/esm.git"
]
 

[dependency-groups]
dev = [
    "pytest",
]

[project.urls]
Homepage = "https://github.com/apple/ml-simplefold"

# Optional: create a CLI command `simplefold` that calls simplefold/cli.py:main()
[project.scripts]
simplefold = "simplefold.cli:main"

# Tell hatchling where your packages live when using src layout:
[tool.hatch.build.targets.wheel]
packages = ["src/simplefold"]


[tool.pytest.ini_options]
addopts = "--pdbcls=IPython.terminal.debugger:TerminalPdb -vv --pdb" # Custom debugger ipdb, -vv for super verbose (logs if each tests passes and logs full error, --pdb to activate debugger).
testpaths = [
    "test",
]